+++ extensions/sha2/rvr/c/rvr_ext_sha2.h
@@ -6,11 +6,11 @@
 struct RvState;
 
 /* SHA-256 compress extension FFI entry point (implemented in Rust). */
-extern void rvr_ext_sha256(RvState* state, uint32_t dst_ptr, uint32_t state_ptr, uint32_t input_ptr,
+extern void rvr_ext_sha256(RvState* state, uint64_t dst_ptr, uint64_t state_ptr, uint64_t input_ptr,
                            uint32_t main_chip_idx, uint32_t block_hasher_chip_idx);
 
 /* SHA-512 compress extension FFI entry point (implemented in Rust). */
-extern void rvr_ext_sha512(RvState* state, uint32_t dst_ptr, uint32_t state_ptr, uint32_t input_ptr,
+extern void rvr_ext_sha512(RvState* state, uint64_t dst_ptr, uint64_t state_ptr, uint64_t input_ptr,
                            uint32_t main_chip_idx, uint32_t block_hasher_chip_idx);
 
 #endif /* RVR_EXT_SHA2_H */
